pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.4.1",
    "ipykernel>=6.29.5",
    "ipython>=8.37.0",
    "polars>=1.31.0",
]
